+++ api/utils.py
@@ -4,7 +4,7 @@ from .models import WorkArrangement
 
 def calculate_employee_monthly_salary(
     work_arrangement, hourly_rate, hours_per_week=40, team_leader=None
-) -> int:
+) -> float:
     """
     Calculate montly salary based on work_arrangement and hourly_rate,
     if employee is team leader, add 10% to salary
@@ -21,4 +21,4 @@ def calculate_employee_monthly_salary(
         # add 10% to salary
         salary = salary + (salary * 0.1)
 
-    return salary
+    return float(salary)
